+++ modules/gacha_log/const.py
@@ -1,0 +1,17 @@
+from genshin.models import BannerType
+
+PAIMONMOE_VERSION = 3
+PM2UIGF_VERSION = 1
+PM2UIGF_NAME = "paimon_moe_to_uigf"
+UIGF_VERSION = "v2.2"
+
+
+GACHA_TYPE_LIST = {
+    BannerType.NOVICE: "新手祈愿",
+    BannerType.PERMANENT: "常驻祈愿",
+    BannerType.WEAPON: "武器祈愿",
+    BannerType.CHARACTER1: "角色祈愿",
+    BannerType.CHARACTER2: "角色祈愿",
+}
+
+
