+++ modules/gacha_log/error.py
@@ -1,0 +1,22 @@
+class GachaLogException(Exception):
+    pass
+
+
+class GachaLogFileError(GachaLogException):
+    pass
+
+
+class GachaLogNotFound(GachaLogException):
+    pass
+
+
+class GachaLogAccountNotFound(GachaLogException):
+    pass
+
+
+class GachaLogInvalidAuthkey(GachaLogException):
+    pass
+
+
+class PaimonMoeGachaLogFileError(GachaLogFileError):
+    pass
