+++ plugins/genshin/gacha/gacha_log.py
@@ -1,13 +1,9 @@
 import json
-from datetime import datetime
-from enum import Enum
 from io import BytesIO
 from os import sep
 
 import genshin
 from genshin.models import BannerType
-from modules.apihelper.gacha_log import GachaLog as GachaLogService
-from openpyxl import load_workbook
 from telegram import Update, User, Message, Document, InlineKeyboardButton, InlineKeyboardMarkup
 from telegram.constants import ChatAction
 from telegram.ext import CallbackContext, CommandHandler, MessageHandler, filters, ConversationHandler
@@ -21,7 +17,9 @@ from core.template import TemplateService
 from core.user import UserService
 from core.user.error import UserNotFoundError
 from modules.apihelper.hyperion import SignIn
+from modules.gacha_log.log import GachaLog
 from utils.bot import get_all_args
+from utils.const import PROJECT_ROOT
 from utils.decorators.admins import bot_admins_rights_check
 from utils.decorators.error import error_callable
 from utils.decorators.restricts import restricts
@@ -29,10 +27,12 @@ from utils.helpers import get_genshin_client
 from utils.log import logger
 from utils.models.base import RegionEnum
 
+GACHA_LOG_PATH = PROJECT_ROOT.joinpath("data", "apihelper", "gacha_log")
+GACHA_LOG_PATH.mkdir(parents=True, exist_ok=True)
 INPUT_URL, INPUT_FILE, CONFIRM_DELETE = range(10100, 10103)
 
 
-class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
+class GachaLogPlugin(Plugin.Conversation, BasePlugin.Conversation):
     """抽卡记录导入/导出/分析"""
 
     def __init__(
@@ -46,20 +46,13 @@ class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
         self.user_service = user_service
         self.assets_service = assets
         self.cookie_service = cookie_service
+        with open(f"resources{sep}json{sep}zh.json", "r") as load_f:
+            self.zh_dict = json.load(load_f)
+        self.gacha_log = GachaLog(GACHA_LOG_PATH)
 
-    @staticmethod
-    def from_url_get_authkey(url: str) -> str:
-        """从 UEL 解析 authkey
-        :param url: URL
-        :return: authkey
-        """
-        try:
-            return url.split("authkey=")[1].split("&")[0]
-        except IndexError:
-            return url
-
-    @staticmethod
-    async def _refresh_user_data(user: User, data: dict = None, authkey: str = None, verify_uid: bool = True) -> str:
+    async def _refresh_user_data(
+        self, user: User, data: dict = None, authkey: str = None, verify_uid: bool = True
+    ) -> str:
         """刷新用户数据
         :param user: 用户
         :param data: 数据
@@ -70,152 +63,14 @@ class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
             logger.debug("尝试获取已绑定的原神账号")
             client = await get_genshin_client(user.id, need_cookie=False)
             if authkey:
-                return await GachaLogService.get_gacha_log_data(user.id, client, authkey)
+                new_num = await self.gacha_log.get_gacha_log_data(user.id, client, authkey)
+                return "更新完成，本次没有新增数据" if new_num == 0 else f"更新完成，本次共新增{new_num}条抽卡记录"
             if data:
-                return await GachaLogService.import_gacha_log_data(user.id, client, data, verify_uid)
+                new_num = await self.gacha_log.import_gacha_log_data(user.id, client, data, verify_uid)
+                return "更新完成，本次没有新增数据" if new_num == 0 else f"更新完成，本次共新增{new_num}条抽卡记录"
         except UserNotFoundError:
             logger.info(f"未查询到用户({user.full_name} {user.id}) 所绑定的账号信息")
             return "派蒙没有找到您所绑定的账号信息，请先私聊派蒙绑定账号"
-
-    @staticmethod
-    def convert_paimonmoe_to_uigf(data: BytesIO) -> dict:
-        """转换 paimone.moe 或 非小酋 导出 xlsx 数据为 UIGF 格式
-        :param data: paimon.moe 导出的 xlsx 数据
-        :return: UIGF 格式数据
-        """
-        PAIMONMOE_VERSION = 3
-        PM2UIGF_VERSION = 1
-        PM2UIGF_NAME = "paimon_moe_to_uigf"
-        UIGF_VERSION = "v2.2"
-
-        with open(f"resources{sep}json{sep}zh.json", "r") as load_f:
-            zh_dict = json.load(load_f)
-
-        class XlsxType(Enum):
-            PAIMONMOE = 1
-            FXQ = 2
-
-        class ItemType(Enum):
-            CHARACTER = "角色"
-            WEAPON = "武器"
-
-        class UIGFGachaType(Enum):
-            BEGINNER = 100
-            STANDARD = 200
-            CHARACTER = 301
-            WEAPON = 302
-
-        class Qiyr:
-            def __init__(
-                self, uigf_gacha_type: UIGFGachaType, item_type: ItemType, name: str, time: datetime, p: int, _id: int
-            ) -> None:
-                self.uigf_gacha_type = uigf_gacha_type
-                self.item_type = item_type
-                self.name = name
-                self.time = time
-                self.rank_type = p
-                self.id = _id
-
-            def qy2_json(self):
-                return {
-                    "gacha_type": self.uigf_gacha_type.value,  # 注意！
-                    "item_id": "",
-                    "count": -1,
-                    "time": self.time.strftime("%Y-%m-%d %H:%M:%S"),
-                    "name": self.name,
-                    "item_type": self.item_type.value,
-                    "rank_type": self.rank_type,
-                    "id": self.id,
-                    "uigf_gacha_type": self.uigf_gacha_type.value,
-                }
-
-        def from_paimon_moe(uigf_gacha_type: UIGFGachaType, item_type: str, name: str, time: str, p: int) -> Qiyr:
-            item_type = ItemType.CHARACTER if type == "Character" else ItemType.WEAPON
-            name = zh_dict[name]
-
-            time = datetime.strptime(time, "%Y-%m-%d %H:%M:%S")
-            return Qiyr(uigf_gacha_type, item_type, name, time, p, 0)
-
-        def from_fxq(uigf_gacha_type: UIGFGachaType, item_type: str, name: str, time: str, p: int, _id: int) -> Qiyr:
-            item_type = ItemType.CHARACTER if type == "角色" else ItemType.WEAPON
-            time = datetime.strptime(time, "%Y-%m-%d %H:%M:%S")
-            return Qiyr(uigf_gacha_type, item_type, name, time, p, _id)
-
-        class uigf:
-            qiyes: list[Qiyr]
-            uid: int
-            export_time: datetime
-            export_app: str = PM2UIGF_NAME
-            export_app_version: str = PM2UIGF_VERSION
-            uigf_version = UIGF_VERSION
-            lang = "zh-cn"
-
-            def __init__(self, qiyes: list[Qiyr], uid: int, export_time: datetime) -> None:
-                self.uid = uid
-                self.qiyes = qiyes
-                self.qiyes.sort(key=lambda x: x.time)
-                if self.qiyes[0].id == 0:  # 如果是从paimon.moe导入的，那么就给id赋值
-                    for index, _ in enumerate(self.qiyes):
-                        self.qiyes[index].id = index + 1
-                    self.export_time = export_time
-                self.export_time = export_time
-
-            def export_json(self) -> dict:
-                json_d = {
-                    "info": {
-                        "uid": self.uid,
-                        "lang": self.lang,
-                        "export_time": self.export_time.strftime("%Y-%m-%d %H:%M:%S"),
-                        "export_timestamp": self.export_time.timestamp(),
-                        "export_app": self.export_app,
-                        "export_app_version": self.export_app_version,
-                        "uigf_version": self.uigf_version,
-                    },
-                    "list": [],
-                }
-                for qiye in self.qiyes:
-                    json_d["list"].append(qiye.qy2_json())
-                return json_d
-
-        wb = load_workbook(data)
-
-        xlsx_type = XlsxType.PAIMONMOE if len(wb.worksheets) == 6 else XlsxType.FXQ  # 判断是paimon.moe还是非小酋导出的
-
-        paimonmoe_sheets = {
-            UIGFGachaType.BEGINNER: "Beginners' Wish",
-            UIGFGachaType.STANDARD: "Standard",
-            UIGFGachaType.CHARACTER: "Character Event",
-            UIGFGachaType.WEAPON: "Weapon Event",
-        }
-        fxq_sheets = {
-            UIGFGachaType.BEGINNER: "新手祈愿",
-            UIGFGachaType.STANDARD: "常驻祈愿",
-            UIGFGachaType.CHARACTER: "角色活动祈愿",
-            UIGFGachaType.WEAPON: "武器活动祈愿",
-        }
-        qiyes = []
-        if xlsx_type == XlsxType.PAIMONMOE:
-            ws = wb["Information"]
-            if ws["B2"].value != PAIMONMOE_VERSION:
-                raise Exception("PaimonMoe version not supported")
-            export_time = datetime.strptime(ws["B3"].value, "%Y-%m-%d %H:%M:%S")
-            for gacha_type in paimonmoe_sheets:
-                ws = wb[paimonmoe_sheets[gacha_type]]
-                for row in ws.iter_rows(min_row=2, values_only=True):
-                    if row[0] is None:
-                        break
-                    qiyes.append(from_paimon_moe(gacha_type, row[0], row[1], row[2], row[3]))
-        else:
-            export_time = datetime.now()
-            for gacha_type in fxq_sheets:
-                ws = wb[fxq_sheets[gacha_type]]
-                for row in ws.iter_rows(min_row=2, values_only=True):
-                    if row[0] is None:
-                        break
-                    qiyes.append(from_fxq(gacha_type, row[2], row[1], row[0], row[3], row[6]))
-
-        u = uigf(qiyes, 0, export_time)
-        return u.export_json()
 
     async def import_from_file(self, user: User, message: Message, document: Document = None) -> None:
         if not document:
@@ -224,11 +79,13 @@ class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
         if document.file_name.endswith(".xlsx"):
             file_type = "xlsx"
         elif document.file_name.endswith(".json"):
-            file_type = "json"
+            file_type = "../../../resources/json"
         else:
             await message.reply_text("文件格式错误，请发送符合 UIGF 标准的 json 格式的抽卡记录文件或者 paimon.moe、非小酋导出的 xlsx 格式的抽卡记录文件")
+            return
         if document.file_size > 2 * 1024 * 1024:
             await message.reply_text("文件过大，请发送小于 2 MB 的文件")
+            return
         try:
             data = BytesIO()
             await (await document.get_file()).download(out=data)
@@ -237,7 +94,7 @@ class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
                 data = data.getvalue().decode("utf-8")
                 data = json.loads(data)
             else:
-                data = self.convert_paimonmoe_to_uigf(data)
+                data = self.gacha_log.convert_xlsx_to_uigf(data, self.zh_dict)
         except UnicodeDecodeError:
             await message.reply_text("文件解析失败，请检查文件编码是否正确或符合 UIGF 标准")
             return
@@ -346,7 +203,7 @@ class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
         except UserNotFoundError:
             await message.reply_text("你还没有导入抽卡记录哦~")
             return ConversationHandler.END
-        _, status = await GachaLogService.load_history_info(str(user.id), str(client.uid), only_status=True)
+        _, status = await self.gacha_log.load_history_info(str(user.id), str(client.uid), only_status=True)
         if not status:
             await message.reply_text("你还没有导入抽卡记录哦~")
             return ConversationHandler.END
@@ -361,7 +218,7 @@ class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
         message = update.effective_message
         user = update.effective_user
         if message.text == "确定":
-            status = await GachaLogService.remove_history_info(str(user.id), str(context.chat_data["uid"]))
+            status = await self.gacha_log.remove_history_info(str(user.id), str(context.chat_data["uid"]))
             await message.reply_text("抽卡记录已删除" if status else "抽卡记录删除失败")
             return ConversationHandler.END
         await message.reply_text("已取消")
@@ -380,11 +237,11 @@ class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
             if cid < 0:
                 raise ValueError("Invalid cid")
             client = await get_genshin_client(cid, need_cookie=False)
-            _, status = await GachaLogService.load_history_info(str(cid), str(client.uid), only_status=True)
+            _, status = await self.gacha_log.load_history_info(str(cid), str(client.uid), only_status=True)
             if not status:
                 await message.reply_text("该用户还没有导入抽卡记录")
                 return
-            status = await GachaLogService.remove_history_info(str(cid), str(client.uid))
+            status = await self.gacha_log.remove_history_info(str(cid), str(client.uid))
             await message.reply_text("抽卡记录已强制删除" if status else "抽卡记录删除失败")
         except UserNotFoundError:
             await message.reply_text("该用户暂未绑定账号")
@@ -402,12 +259,9 @@ class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
         try:
             client = await get_genshin_client(user.id, need_cookie=False)
             await message.reply_chat_action(ChatAction.TYPING)
-            state, text, path = await GachaLogService.gacha_log_to_uigf(str(user.id), str(client.uid))
-            if state:
-                await message.reply_chat_action(ChatAction.UPLOAD_DOCUMENT)
-                await message.reply_document(document=open(path, "rb+"), caption="抽卡记录导出文件")
-            else:
-                await message.reply_text(text)
+            path = await self.gacha_log.gacha_log_to_uigf(str(user.id), str(client.uid))
+            await message.reply_chat_action(ChatAction.UPLOAD_DOCUMENT)
+            await message.reply_document(document=open(path, "rb+"), caption="抽卡记录导出文件")
         except UserNotFoundError:
             logger.info(f"未查询到用户({user.full_name} {user.id}) 所绑定的账号信息")
             if filters.ChatType.GROUPS.filter(message):
@@ -433,7 +287,7 @@ class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
         try:
             client = await get_genshin_client(user.id, need_cookie=False)
             await message.reply_chat_action(ChatAction.TYPING)
-            data = await GachaLogService.get_analysis(user.id, client, pool_type, self.assets_service)
+            data = await self.gacha_log.get_analysis(user.id, client, pool_type, self.assets_service)
             if isinstance(data, str):
                 reply_message = await message.reply_text(data)
                 if filters.ChatType.GROUPS.filter(message):
@@ -476,7 +330,7 @@ class GachaLog(Plugin.Conversation, BasePlugin.Conversation):
             client = await get_genshin_client(user.id, need_cookie=False)
             group = filters.ChatType.GROUPS.filter(message)
             await message.reply_chat_action(ChatAction.TYPING)
-            data = await GachaLogService.get_pool_analysis(user.id, client, pool_type, self.assets_service, group)
+            data = await self.gacha_log.get_pool_analysis(user.id, client, pool_type, self.assets_service, group)
             if isinstance(data, str):
                 reply_message = await message.reply_text(data)
                 if filters.ChatType.GROUPS.filter(message):
